// src/shared/api/generated/models/CoworkingFloorPlaceResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CRM API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CoworkingAccountBookingResponse } from './CoworkingAccountBookingResponse';
import {
    CoworkingAccountBookingResponseFromJSON,
    CoworkingAccountBookingResponseFromJSONTyped,
    CoworkingAccountBookingResponseToJSON,
    CoworkingAccountBookingResponseToJSONTyped,
} from './CoworkingAccountBookingResponse';

/**
 * 
 * @export
 * @interface CoworkingFloorPlaceResponse
 */
export interface CoworkingFloorPlaceResponse {
    /**
     * 
     * @type {string}
     * @memberof CoworkingFloorPlaceResponse
     */
    placeId: string;
    /**
     * 
     * @type {number}
     * @memberof CoworkingFloorPlaceResponse
     */
    latitude: number;
    /**
     * 
     * @type {number}
     * @memberof CoworkingFloorPlaceResponse
     */
    longitude: number;
    /**
     * 
     * @type {number}
     * @memberof CoworkingFloorPlaceResponse
     */
    price: number;
    /**
     * 
     * @type {Array<CoworkingAccountBookingResponse>}
     * @memberof CoworkingFloorPlaceResponse
     */
    bookings: Array<CoworkingAccountBookingResponse>;
}

/**
 * Check if a given object implements the CoworkingFloorPlaceResponse interface.
 */
export function instanceOfCoworkingFloorPlaceResponse(value: object): value is CoworkingFloorPlaceResponse {
    if (!('placeId' in value) || value['placeId'] === undefined) return false;
    if (!('latitude' in value) || value['latitude'] === undefined) return false;
    if (!('longitude' in value) || value['longitude'] === undefined) return false;
    if (!('price' in value) || value['price'] === undefined) return false;
    if (!('bookings' in value) || value['bookings'] === undefined) return false;
    return true;
}

export function CoworkingFloorPlaceResponseFromJSON(json: any): CoworkingFloorPlaceResponse {
    return CoworkingFloorPlaceResponseFromJSONTyped(json, false);
}

export function CoworkingFloorPlaceResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CoworkingFloorPlaceResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'placeId': json['place_id'],
        'latitude': json['latitude'],
        'longitude': json['longitude'],
        'price': json['price'],
        'bookings': ((json['bookings'] as Array<any>).map(CoworkingAccountBookingResponseFromJSON)),
    };
}

export function CoworkingFloorPlaceResponseToJSON(json: any): CoworkingFloorPlaceResponse {
    return CoworkingFloorPlaceResponseToJSONTyped(json, false);
}

export function CoworkingFloorPlaceResponseToJSONTyped(value?: CoworkingFloorPlaceResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'place_id': value['placeId'],
        'latitude': value['latitude'],
        'longitude': value['longitude'],
        'price': value['price'],
        'bookings': ((value['bookings'] as Array<any>).map(CoworkingAccountBookingResponseToJSON)),
    };
}